Refactor extracting and renaming ResponseDocumentSummary
